- moved editor component events to separate component to reduce amount of rerendering
